feat: 정당, 의원 팔로우시 팔로워 숫자 즉각 반영되도록 수정
Related to: #167
