+++ app/party/[id]/components/PartyDetail/PartyDetail.tsx
@@ -1,6 +1,5 @@
 import { useQueryClient } from '@tanstack/react-query';
 import { Button } from '@nextui-org/button';
-import Link from 'next/link';
 import Image from 'next/image';
 import { Link as NextUILink } from '@nextui-org/link';
 import { Card } from '@nextui-org/card';
@@ -15,7 +14,7 @@ import {
   PARTY_POLISY_COMMITTEE_CHAIRMAN,
 } from '@/constants/party';
 import { useGetPartyDetail } from '../../apis';
-import FollowButton from './FollowButton';
+import FollowBoard from './FollowBoard';
 
 export default async function PartyDetail({ partyId }: { partyId: number }) {
   const queryClient = useQueryClient();
@@ -36,8 +35,8 @@ export default async function PartyDetail({ partyId }: { partyId: number }) {
   const seatRatio = ((100 * whole_representative_count) / 298).toFixed(2);
 
   return (
-    <section className="flex flex-col items-center mx-5 mt-5 gap-7">
-      <Card shadow="none" className="flex flex-col items-center w-full pt-1 gap-5 dark:bg-dark-b lg:dark:bg-dark-pb">
+    <section className="flex flex-col items-center mx-5 mt-5 gap-7 lg:w-[320px]">
+      <Card shadow="none" className="flex flex-col items-center w-full gap-5 pt-1 dark:bg-dark-b lg:dark:bg-dark-pb">
         <div
           className={`shadow-lg rounded-full w-[130px] h-[130px] flex justify-center items-center border dark:bg-white ${party_name}`}>
           <Image
@@ -63,42 +62,22 @@ export default async function PartyDetail({ partyId }: { partyId: number }) {
             지역구 {district_congressman_count}석, 비례대표 {proportional_congressman_count}석
           </p>
         </div>
-
-        <FollowButton id={partyId} followed={followed} />
       </Card>
 
       <Divider className="dark:bg-dark-l" />
 
-      <section className="grid grid-cols-3 gap-10">
-        <div className="flex flex-col items-center gap-[10px]">
-          <p className="text-2xl font-semibold">{follow_count}</p>
-          <p className="text-sm font-medium text-gray-2 dark:text-gray-3">팔로워</p>
-        </div>
-        <div className="flex flex-col items-center gap-[10px]">
-          {representative_bill_count ? (
-            <Link href={`/party/${party_id}/bill`} className="text-2xl font-semibold">
-              {representative_bill_count}
-            </Link>
-          ) : (
-            <p className="text-2xl font-semibold"> {representative_bill_count}</p>
-          )}
-          <p className="text-sm font-medium text-gray-2 dark:text-gray-3">대표발의법안</p>
-        </div>
-        <div className="flex flex-col items-center gap-[10px]">
-          {public_bill_count ? (
-            <Link href={public_bill_count ? `/party/${party_id}/bill` : ''} className="text-2xl font-semibold">
-              <p className="text-2xl font-semibold">{public_bill_count}</p>
-            </Link>
-          ) : (
-            <p className="text-2xl font-semibold"> {public_bill_count}</p>
-          )}
-          <p className="text-sm font-medium text-gray-2 dark:text-gray-3">공동발의법안</p>
-        </div>
-      </section>
+      <FollowBoard
+        id={partyId}
+        followed={followed}
+        follow_count={follow_count}
+        representative_bill_count={representative_bill_count}
+        public_bill_count={public_bill_count}
+        party_id={party_id}
+      />
 
       <Divider className="dark:bg-dark-l" />
 
-      <div className="w-full grid grid-cols-4 justify-items-center ">
+      <div className="grid w-full grid-cols-4 justify-items-center ">
         <div className="flex flex-col items-center">
           <p className="text-sm text-gray-3 dark:text-gray-2">당대표</p>
           <p className="font-medium">{PARTY_LEADER[party_name as keyof typeof PARTY_NAME_KO] || '없음'}</p>
